requirements.txt
asgiref==3.8.1
Django==5.1.7
django-tinymce==4.1.0
gunicorn==23.0.0
narwhals==1.32.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pep8==1.7.1
pillow==11.1.0
plotly==6.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
